Quals/forum: add description, separate from introduction #196462
There are now separate fields:

'intro' in the database, "Forum description (shown beside links to this
forum, if the option below is on)" in the settings form. This gets
displayed in the study planner (an on the OSEP forums page if
applicable). if the "Display description on main page" option is on.

'introduction' in the database, "Forum introduction (shown above the
list of discussions)" in the settings. This is the text that was already
present in the forum. It has just been moved to a new place so the
standard Moodle Display description on main page feature can work for
the other text.

The confusingly similar names are because Moodle core requires the
description field to be called intro in the database. Hopefully the
comment there is sufficient to explain what is going on.

There is upgrade code to move the existing forum introducitons into the
new field. There is also code in restore, so that the data will be
updated (including files) on restore.

This commit also fixes some of the behat tests that were failing.